requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers>=3.0.0
huggingface-hub>=0.23.0
faiss-cpu
google-generativeai
beautifulsoup4
pymupdf
requests
python-multipart
selenium
webdriver-manager
numpy
